INFRA-43 Adding open source AWS Secrets Manager Dropwizard Authenticator
This PR changes our fork of the original com.washingtonpost Authenticator I wrote years ago that dependend on an encrypted, classpath-accessible allowed-peers.properties and swaps that assumption out for support for AWS Secrets Manager.

This change was tested on localhost with a SNAPSHOT integration of this JAR with the echo-service, using 2 different echo-service Secrets Manager coordinates (service/alpha/echo/basic_auth/admin and service/alpha/echo/basic_auth/general).  I tested the service works with JWDP debugger walk through in the echo-service's docker-compose function and with the ./regression.sh script in that service